+++ memory2/boying/boying-mbg/src/main/java/com/tongji/boying/model/BoyingOrderExample.java
@@ -1,6 +1,5 @@
 package com.tongji.boying.model;
 
-import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
@@ -596,52 +595,52 @@ public class BoyingOrderExample {
             return (Criteria) this;
         }
 
-        public Criteria andMoneyEqualTo(BigDecimal value) {
+        public Criteria andMoneyEqualTo(Double value) {
             addCriterion("MONEY =", value, "money");
             return (Criteria) this;
         }
 
-        public Criteria andMoneyNotEqualTo(BigDecimal value) {
+        public Criteria andMoneyNotEqualTo(Double value) {
             addCriterion("MONEY <>", value, "money");
             return (Criteria) this;
         }
 
-        public Criteria andMoneyGreaterThan(BigDecimal value) {
+        public Criteria andMoneyGreaterThan(Double value) {
             addCriterion("MONEY >", value, "money");
             return (Criteria) this;
         }
 
-        public Criteria andMoneyGreaterThanOrEqualTo(BigDecimal value) {
+        public Criteria andMoneyGreaterThanOrEqualTo(Double value) {
             addCriterion("MONEY >=", value, "money");
             return (Criteria) this;
         }
 
-        public Criteria andMoneyLessThan(BigDecimal value) {
+        public Criteria andMoneyLessThan(Double value) {
             addCriterion("MONEY <", value, "money");
             return (Criteria) this;
         }
 
-        public Criteria andMoneyLessThanOrEqualTo(BigDecimal value) {
+        public Criteria andMoneyLessThanOrEqualTo(Double value) {
             addCriterion("MONEY <=", value, "money");
             return (Criteria) this;
         }
 
-        public Criteria andMoneyIn(List<BigDecimal> values) {
+        public Criteria andMoneyIn(List<Double> values) {
             addCriterion("MONEY in", values, "money");
             return (Criteria) this;
         }
 
-        public Criteria andMoneyNotIn(List<BigDecimal> values) {
+        public Criteria andMoneyNotIn(List<Double> values) {
             addCriterion("MONEY not in", values, "money");
             return (Criteria) this;
         }
 
-        public Criteria andMoneyBetween(BigDecimal value1, BigDecimal value2) {
+        public Criteria andMoneyBetween(Double value1, Double value2) {
             addCriterion("MONEY between", value1, value2, "money");
             return (Criteria) this;
         }
 
-        public Criteria andMoneyNotBetween(BigDecimal value1, BigDecimal value2) {
+        public Criteria andMoneyNotBetween(Double value1, Double value2) {
             addCriterion("MONEY not between", value1, value2, "money");
             return (Criteria) this;
         }
